requirements.txt
streamlit==1.26.0
fpdf2==2.7.8




